+++ src/main/java/ir/maktab/model/repository/AdminRepository.java
@@ -4,7 +4,7 @@ import ir.maktab.model.entity.Admin;
 import org.springframework.data.repository.Repository;
 
 @org.springframework.stereotype.Repository
-public interface AdminRepository  extends Repository<Admin,Integer> {
+public interface AdminRepository  extends Repository<Admin,Long> {
 
     Admin save(Admin admin);
 }
