+++ src/main/java/ir/maktab/model/repository/MultipleChoiceQuestionRepository.java
@@ -4,7 +4,7 @@ import ir.maktab.model.entity.MultipleChoiceQuestion;
 import org.springframework.data.repository.Repository;
 
 @org.springframework.stereotype.Repository
-public interface MultipleChoiceQuestionRepository extends Repository<MultipleChoiceQuestion, Integer> {
+public interface MultipleChoiceQuestionRepository extends Repository<MultipleChoiceQuestion, Long> {
 
     MultipleChoiceQuestion save(MultipleChoiceQuestion multipleChoiceQuestion);
 }
