+++ src/main/java/ir/maktab/model/repository/StudentRepository.java
@@ -4,7 +4,7 @@ import ir.maktab.model.entity.Student;
 import org.springframework.data.repository.Repository;
 
 @org.springframework.stereotype.Repository
-public interface StudentRepository extends Repository<Student,Integer> {
+public interface StudentRepository extends Repository<Student,Long> {
 
     Student save(Student student);
 }
